pyproject.toml
[project]
name = "try-on-api"
version = "0.1.0"
description = "Api for clothes try on for sinpie"
authors = [
    { name = "ranjanmangla1", email = "ranjanmangla1@gmail.com" }
]
dependencies = [
    "fastapi>=0.112.0",
    "pydantic>=2.8.2",
    "uvicorn>=0.30.6",
    "requests>=2.32.3",
    "loguru>=0.7.2",
    "pydantic-settings>=2.4.0",
    "python-multipart>=0.0.9",
    "pillow>=10.4.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "pre-commit>=3.8.0",
    "ruff>=0.5.7",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/try_on_api"]


[tool.rye.scripts]
dev = "python -b server.py"
